x/reward/types/message_create_voucher_test.go
package types

import (
	"testing"

	"titan/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateVoucher_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateVoucher
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateVoucher{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateVoucher{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
